pyproject.toml
[tool.poetry]
name = "codavatar"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
black = "^23.7.0"
python-multipart = "^0.0.6"
fastapi = {extras = ["all"], version = "^0.100.1"}
uvicorn = {extras = ["all"], version = "^0.23.2"}
psycopg2-binary = "^2.9.6"
sqlalchemy = "^2.0.19"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
